pyproject.toml
[project]
name = "transformerllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "tokenizers>=0.22.2",
]
